Doodlebox
A modification to the Mastodon interface that allows you to select a
doodle icon in addition to uploading a media file from your computer.
The doodle interface allows you to draw using your mouse, finger, or
pointing device in several colors. It was originally developed by
Ondřej Hruška <ondra@ondrovo.com> for use on the GlitchSoc fork.